Inactive_Customers/requirements.txt
# Core Google Cloud dependencies
google-api-core==2.19.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
db-dtypes==1.2.1

# Machine Learning
xgboost==2.1.0
scikit-learn==1.5.1
imbalanced-learn==0.12.3
scipy==1.13.1

# Data Processing
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
pandas-gbq==0.23.1

# Visualization
matplotlib==3.9.0
seaborn==0.13.2

# Utilities
python-dateutil==2.9.0
pytz==2024.1
tqdm==4.66.4
joblib==1.4.2

# Development
ipython==8.26.0
ipykernel==6.29.4
jupyter==1.0.0
notebook==7.2.1
